pyproject.toml
[project]
name = "chemprop-transformer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
    "pyspark>=4.0.0",
    "pyyaml>=6.0.2",
    "rdkit>=2025.3.3",
    "rotary-embedding-torch>=0.8.7",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "selfies>=2.2.0",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "werkzeug>=3.1.3",
    "x-transformers>=2.4.9",
    "psutil==7.0.0",
    "termcolor==3.1.0",
    "entmax>=1.3",
    "gradnorm-pytorch>=0.1.1",
    "watchdog>=6.0.0",
    "timm>=1.0.19",
]
